fix debug response when a response var is present
Previously, the entire response returned by cue was also the crossplane
response so it made sense to dump it as-is such that users could
copy the request and response objects unmodified into their unit tests.

Now that we support extracting a variable from the response, we need to
replay it back with that variable such that copying the debug response output
unmodified yields a correct test. That is, we do the same processing for
responses as we do for requests.

Signed-off-by: gotwarlost <krishnan.anantheswaran@elastic.co>